Extract setProjectConfigForErrors in a helper
